+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 #Changelog
 
 
+### 0.2.5 (2016-06-26)
+
+- remove ctrl+click feature to activate all language tabs of the same language (core feature from ProcessWire 3.023)
+- FireFox fix for clearing filterbox
+
+
+
 ### 0.2.3 (2016-06-25)
 
 - new submodule NoAnims: disable all admin animations